Add comments to example programs, fix blueprints.cpp - fixed OR blueprint misalign - example progams now have left aligned comments of their instruction # - batch_test.sh now has a regenerate only flag

+++ testing/batch_test.sh
@@ -23,13 +23,19 @@ if [[ ! -d ${testing_dir}/example_programs ]]; then
     mkdir ${testing_dir}/example_programs
 fi
 
-# only try to delete example_programs if they already exist
-if [[ -f ${testing_dir}/example_programs/* ]]; then
-    for i in ${testing_dir}/example_programs/*; do
-        rm $i
-    done
+# only regenerate files
+if [[ $# -eq 1 ]]; then
+    if [[ $1 == regenerate ]]; then
+        # only try to delete example_programs if they already exist
+        if [[ -f ${testing_dir}/example_programs/* ]]; then
+            for i in ${testing_dir}/example_programs/*; do
+                rm $i
+            done
+        fi
+        python3 ${test_gen}
+    fi
+    exit
 fi
-python3 ${test_gen}
 
 # test valid, then test error
 for i in ${testing_dir}/example_programs/gib_valid_*; do
